+++ examples/Rules-method-nextbest-NextBestTDsamples.R
@@ -26,8 +26,8 @@ my_samples <- mcmc(my_data, my_model, my_options)
 tds_next_best <- NextBestTDsamples(
   prob_target_drt = 0.35,
   prob_target_eot = 0.3,
-  derive = function(dose_samples) { # nolintr
-    as.numeric(quantile(dose_samples, probs = 0.3))
+  derive = function(samples) {
+    as.numeric(quantile(samples, probs = 0.3))
   }
 )