+++ tests/testthat/test-Rules-class.R
@@ -9,16 +9,11 @@ test_that(".NextBestMTD works as expected", {
 
 test_that("NextBestMTD object can be created with user constructor", {
   result <- expect_silent(
-    NextBestMTD(
-      target = 0.4,
-      derive = function(mtd_samples) {
-        mean(mtd_samples)
-      }
-    )
+    NextBestMTD(0.4, function(mtd_samples) mean(mtd_samples))
   )
   expect_valid(result, "NextBestMTD")
-  expect_equal(result@target, 0.4)
-  expect_equal(result@derive(c(1:5)), 3) # nolintr
+  expect_identical(result@target, 0.4)
+  expect_identical(result@derive(c(1:5)), 3) # nolintr
 })
 
 ## NextBestNCRM ----
@@ -30,16 +25,12 @@ test_that(".NextBestNCRM works as expected", {
 
 test_that("NextBestNCRM object can be created with user constructor", {
   result <- expect_silent(
-    NextBestNCRM(
-      target = c(0.5, 0.8),
-      overdose = c(0.4, 1),
-      max_overdose_prob = 0.3
-    )
+    NextBestNCRM(c(0.5, 0.8), c(0.4, 1), 0.3)
   )
   expect_valid(result, "NextBestNCRM")
-  expect_equal(result@target, c(0.5, 0.8))
-  expect_equal(result@overdose, c(0.4, 1))
-  expect_equal(result@max_overdose_prob, 0.3)
+  expect_identical(result@target, c(0.5, 0.8))
+  expect_identical(result@overdose, c(0.4, 1))
+  expect_identical(result@max_overdose_prob, 0.3)
 })
 
 ## NextBestNCRMLoss ----
@@ -51,19 +42,14 @@ test_that(".NextBestNCRMLoss works as expected", {
 
 test_that("NextBestNCRMLoss object can be created with user constructor", {
   result <- expect_silent(
-    NextBestNCRMLoss(
-      target = c(0.2, 0.35),
-      overdose = c(0.35, 0.6),
-      unacceptable = c(0.6, 1),
-      max_overdose_prob = 0.25,
-      losses = c(1, 0, 1, 2)
-    )
+    NextBestNCRMLoss(c(0.2, 0.35), c(0.35, 0.6), c(0.6, 1), 0.25, c(1, 0, 1, 2))
   )
-  expect_equal(result@target, c(0.2, 0.35))
-  expect_equal(result@overdose, c(0.35, 0.6))
-  expect_equal(result@unacceptable, c(0.6, 1))
-  expect_equal(result@max_overdose_prob, 0.25)
-  expect_equal(result@losses, c(1, 0, 1, 2))
+  expect_valid(result, "NextBestNCRMLoss")
+  expect_identical(result@target, c(0.2, 0.35))
+  expect_identical(result@overdose, c(0.35, 0.6))
+  expect_identical(result@unacceptable, c(0.6, 1))
+  expect_identical(result@max_overdose_prob, 0.25)
+  expect_identical(result@losses, c(1, 0, 1, 2))
 })
 
 ## NextBestThreePlusThree ----
@@ -87,34 +73,24 @@ test_that(".NextBestDualEndpoint works as expected", {
 
 test_that("NextBestDualEndpoint object can be created with user constructor", {
   result <- expect_silent(
-    NextBestDualEndpoint(
-      target = c(0.5, 0.8),
-      overdose = c(0.4, 1),
-      max_overdose_prob = 0.3
-    )
+    NextBestDualEndpoint(c(0.5, 0.8), c(0.4, 1), 0.3)
   )
   expect_valid(result, "NextBestDualEndpoint")
-  expect_equal(result@target, c(0.5, 0.8))
-  expect_equal(result@overdose, c(0.4, 1))
-  expect_equal(result@target_relative, TRUE)
-  expect_equal(result@target_thresh, 0.01)
+  expect_identical(result@target, c(0.5, 0.8))
+  expect_identical(result@overdose, c(0.4, 1))
+  expect_identical(result@target_relative, TRUE)
+  expect_identical(result@target_thresh, 0.01)
 })
 
 test_that("NextBestDualEndpoint object can be created with user constructor 2", {
   result <- expect_silent(
-    NextBestDualEndpoint(
-      target = c(0.5, 0.8),
-      target_relative = FALSE,
-      overdose = c(0.4, 1),
-      max_overdose_prob = 0.3,
-      target_thresh = 0.05
-    )
+    NextBestDualEndpoint(c(0.5, 0.8), FALSE, c(0.4, 1), 0.3, 0.05)
   )
   expect_valid(result, "NextBestDualEndpoint")
-  expect_equal(result@target, c(0.5, 0.8))
-  expect_equal(result@overdose, c(0.4, 1))
-  expect_equal(result@target_relative, FALSE)
-  expect_equal(result@target_thresh, 0.05)
+  expect_identical(result@target, c(0.5, 0.8))
+  expect_identical(result@overdose, c(0.4, 1))
+  expect_identical(result@target_relative, FALSE)
+  expect_identical(result@target_thresh, 0.05)
 })
 
 ## NextBestMinDist ----
@@ -126,12 +102,10 @@ test_that("NextBestMinDist works as expected", {
 
 test_that("NextBestMinDist object can be created with user constructor", {
   result <- expect_silent(
-    NextBestMinDist(
-      target = 0.3
-    )
+    NextBestMinDist(0.3)
   )
   expect_valid(result, "NextBestMinDist")
-  expect_equal(result@target, 0.3)
+  expect_identical(result@target, 0.3)
 })
 
 ## NextBestInfTheory ----
@@ -143,14 +117,11 @@ test_that(".NextBestInfTheory works as expected", {
 
 test_that("NextBestInfTheory object can be created with user constructor", {
   result <- expect_silent(
-    NextBestInfTheory(
-      target = 0.4,
-      asymmetry = 1.5
-    )
+    NextBestInfTheory(0.4, 1.5)
   )
   expect_valid(result, "NextBestInfTheory")
-  expect_equal(result@target, 0.4)
-  expect_equal(result@asymmetry, 1.5)
+  expect_identical(result@target, 0.4)
+  expect_identical(result@asymmetry, 1.5)
 })
 
 ## NextBestTD ----
@@ -165,8 +136,8 @@ test_that("NextBestTD object can be created with user constructor", {
     NextBestTD(0.4, 0.35)
   )
   expect_valid(result, "NextBestTD")
-  expect_equal(result@prob_target_drt, 0.4)
-  expect_equal(result@prob_target_eot, 0.35)
+  expect_identical(result@prob_target_drt, 0.4)
+  expect_identical(result@prob_target_eot, 0.35)
 })
 
 ## NextBestTDsamples ----
@@ -181,9 +152,9 @@ test_that("NextBestTDsamples object can be created with user constructor", {
     NextBestTDsamples(0.4, 0.35, function(x) mean(x))
   )
   expect_valid(result, "NextBestTDsamples")
-  expect_equal(result@prob_target_drt, 0.4)
-  expect_equal(result@prob_target_eot, 0.35)
-  expect_equal(result@derive(c(1:5)), 3) # nolintr
+  expect_identical(result@prob_target_drt, 0.4)
+  expect_identical(result@prob_target_eot, 0.35)
+  expect_identical(result@derive(c(1:5)), 3) # nolintr
 })
 
 ## NextBestMaxGain ----
@@ -198,8 +169,26 @@ test_that("NextBestMaxGain object can be created with user constructor", {
     NextBestMaxGain(0.4, 0.35)
   )
   expect_valid(result, "NextBestMaxGain")
-  expect_equal(result@prob_target_drt, 0.4)
-  expect_equal(result@prob_target_eot, 0.35)
+  expect_identical(result@prob_target_drt, 0.4)
+  expect_identical(result@prob_target_eot, 0.35)
+})
+
+## NextBestMaxGainSamples ----
+
+test_that(".NextBestMaxGainSamples works as expected", {
+  result <- expect_silent(.NextBestMaxGainSamples())
+  expect_valid(result, "NextBestMaxGainSamples")
+})
+
+test_that("NextBestMaxGainSamples object can be created with user constructor", {
+  result <- expect_silent(
+    NextBestMaxGainSamples(0.4, 0.35, function(x) mean(x), function(x) sum(x))
+  )
+  expect_valid(result, "NextBestMaxGainSamples")
+  expect_identical(result@prob_target_drt, 0.4)
+  expect_identical(result@prob_target_eot, 0.35)
+  expect_identical(result@derive(c(1:5)), 3) # nolintr
+  expect_identical(result@mg_derive(c(1:5)), 15) # nolintr
 })
 
 # Increments ----