+++ NEWS.md
@@ -1,5 +1,6 @@
 # Version 1.0.9000.9065
-* Renamed slots in `NextBestTD`, `NextBestTDsamples`, and `NextBestMaxGain` classes.
+* Renamed slots in `NextBestTD`, `NextBestTDsamples`, `NextBestMaxGain` and
+  `NextBestMaxGainSamples` classes.
 * Removed `Model` class.
 * Renamed `SIM` argument of `nextBest` methods to `in_sim`.
 * Renamed the elements of the list returned by `nextBest` methods for `NextBestMaxGain`